refactor: extract all changes except moto claim from feat/dom/moto-drop
