+++ src/containers/flows/ClaimNFT/ClaimNFT.constants.ts
@@ -22,7 +22,6 @@ export const LABELS = {
 export const DOMAINS = {
 	ELIGIBLE_NFT_ROUTE: '/wheels.genesis',
 	CURRENT_CLAIM_ROUTE: '/moto.genesis',
-	WHEELS_DOMAIN_NAME: 'wilder.wheels.genesis',
 };
 
 export const MESSAGES = {
